Set likelihood within the init method of the likelihood
The binary is not anymore the place where all the options are treated. The likelihood now only
accepts a posterior sample json file with all the likelihood options.

Tests between this version and the historical one gave same results for posterior only mode, skymap
only mode and skymap with different cosmology (H0 and Omega_m).